backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specter-backend"
version = "0.1.0"
description = "FastAPI control plane for the SPECTER YC hackathon project"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.52.0",
  "browser-use>=0.12.0",
  "exa-py>=2.6.1",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "google-genai>=1.60.0",
  "langchain-openai>=0.3.0",
  "lmnr>=0.7.40",
  "loguru>=0.7.3",
  "mediapipe>=0.10.32",
  "numpy>=2.4.2",
  "opencv-python-headless>=4.10.0",
  "PicImageSearch>=3.12.11",
  "pillow>=12.1.0",
  "pydantic-settings>=2.10.1",
  "python-multipart>=0.0.20",
  "twscrape>=0.17.0",
  "ultralytics>=8.3.0",
  "uvicorn>=0.35.0",
  "python-telegram-bot>=22.6",
  "hud-python>=0.5.29",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.1",
  "pytest-cov>=6.0",
  "ruff>=0.12.11",
]

[tool.setuptools]
py-modules = ["config", "env_check", "main", "pipeline", "schemas", "tasks"]
packages = ["agents", "capture", "db", "enrichment", "eval", "identification", "observability", "synthesis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-config=pyproject.toml"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["."]
omit = [
  "tests/*",
  ".venv/*",
  "__pycache__/*",
]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
